Move main app root and test root away from each other
