+++ src/common/test/Workspace.spec.ts
@@ -1,6 +1,6 @@
 import EditorGateway from 'gateway/EditorGateway';
 import Workspace from 'common/Workspace';
-import getContainer from 'common/TestAppRoot';
+import getContainer from 'test-tool/TestRoot';
 
 let editorGateway: EditorGateway;
 let workspace: Workspace;
