+++ src/config/test/ConfigFile.spec.ts
@@ -1,7 +1,7 @@
 import ConfigFile from 'config/ConfigFile';
 import EditorGateway from 'gateway/EditorGateway';
 import getContextStub from 'test-tool/helper/getContextStub';
-import getContainer from 'common/TestAppRoot';
+import getContainer from 'test-tool/TestRoot';
 
 function getMockContext(filePath: string) {
   const initialContext = getContextStub();
