+++ src/containerization/test/UncommittedFile.spec.ts
@@ -2,7 +2,7 @@ import ChildProcessGateway from 'gateway/ChildProcessGateway';
 import EditorGateway from 'gateway/EditorGateway';
 import getContextStub from 'test-tool/helper/getContextStub';
 import UncommittedFile from 'containerization/UncommittedFile';
-import getContainer from 'common/TestAppRoot';
+import getContainer from 'test-tool/TestRoot';
 
 function getMockContext(isSuppressed: boolean) {
   const initialContext = getContextStub();
